v1.2.0
Add switch and outlet accessory.
Add for all accessories json as response type.
