+++ src/index.ts
@@ -17,6 +17,10 @@ const pjson = require('../package.json');
 
 import { md5 } from "./md5";
 
+import { SwitchAccessory }    from "./accessories/SwitchAccessory"
+import { LightbulbAccessory } from "./accessories/LightbulbAccessory"
+import { OutletAccessory }    from "./accessories/OutletAccessory"
+
 let hap: HAP;
 
 export = (api: API) => {
@@ -25,6 +29,9 @@ export = (api: API) => {
 };
 
 class HttpIotAccessory implements AccessoryPlugin {
+
+  serviceToExpose: any;
+  informationService: Service;
 
   private readonly log:    Logging;
   private readonly name:   string;
@@ -39,8 +46,9 @@ class HttpIotAccessory implements AccessoryPlugin {
     Saturation:       0,     // number 0% - 100%
   };
 
-  private readonly accessoryService: Service;
-  private readonly informationService: Service;
+  lightbulbAccessory: LightbulbAccessory;
+  switchAccessory:    SwitchAccessory;
+  outletAccessory:    OutletAccessory;
 
   constructor(log: Logging, config: AccessoryConfig, api: API) {
     this.log    = log;
@@ -48,129 +56,112 @@ class HttpIotAccessory implements AccessoryPlugin {
     this.config = config;
     this.api    = api;
 
+    this.lightbulbAccessory = new LightbulbAccessory(this.log, this.config, this.accStates, this.api);
+    this.switchAccessory    = new SwitchAccessory(this.log, this.config, this.accStates, this.api);
+    this.outletAccessory    = new OutletAccessory(this.log, this.config, this.accStates, this.api);
+
     if ((!this.config.type) || (this.config.type == "lightbulb")) {
 
-      this.lightbulbErrorCheck();
-      this.accessoryService = new hap.Service.Lightbulb(this.name);
+      let lightbulbService: Service = new hap.Service.Lightbulb(this.name);
 
-      if (this.lightbulbCheckOnOff()) {
-        this.accessoryService.getCharacteristic(hap.Characteristic.On)
+      this.lightbulbAccessory.lightbulbErrorCheck();
+
+      if (this.lightbulbAccessory.lightbulbCheckOnOff()) {
+        lightbulbService.getCharacteristic(hap.Characteristic.On)
           .on(CharacteristicEventTypes.GET, async (callback: CharacteristicGetCallback) => {
             callback(undefined, this.accStates.On);
-            this.lightbulbUpdateOn();
+            this.lightbulbAccessory.lightbulbUpdateOn();
           })
           .on(CharacteristicEventTypes.SET, async (value: CharacteristicValue, callback: CharacteristicSetCallback) => {
-            this.lightbulbSetOn(value);
+            this.lightbulbAccessory.lightbulbSetOn(value);
             callback();
           });
       }
 
-      if (this.lightbulbCheckBrightness()) {
-        this.accessoryService.getCharacteristic(hap.Characteristic.Brightness)
+      if (this.lightbulbAccessory.lightbulbCheckBrightness()) {
+        lightbulbService.getCharacteristic(hap.Characteristic.Brightness)
           .on(CharacteristicEventTypes.GET, async (callback: CharacteristicGetCallback) => {
             callback(undefined, this.accStates.Brightness);
-            this.lightbulbUpdateBrightness();
+            this.lightbulbAccessory.lightbulbUpdateBrightness();
           })
           .on(CharacteristicEventTypes.SET, async (value: CharacteristicValue, callback: CharacteristicSetCallback) => {
-            this.lightbulbSetBrightness(value);
+            this.lightbulbAccessory.lightbulbSetBrightness(value);
             callback();
           });
       }
 
-      if (this.lightbulbCheckHue()) {
-        this.accessoryService.getCharacteristic(hap.Characteristic.Hue)
+      if (this.lightbulbAccessory.lightbulbCheckHue()) {
+        lightbulbService.getCharacteristic(hap.Characteristic.Hue)
           .on(CharacteristicEventTypes.GET, async (callback: CharacteristicGetCallback) => {
             callback(undefined, this.accStates.Hue);
-            this.lightbulbUpdateHue();
+            this.lightbulbAccessory.lightbulbUpdateHue();
           })
           .on(CharacteristicEventTypes.SET, async (value: CharacteristicValue, callback: CharacteristicSetCallback) => {
-            this.lightbulbSetHue(value);
+            this.lightbulbAccessory.lightbulbSetHue(value);
             callback();
           });
       }
 
-      if (this.lightbulbCheckColorTemperature()) {
-        this.accessoryService.getCharacteristic(hap.Characteristic.ColorTemperature)
+      if (this.lightbulbAccessory.lightbulbCheckColorTemperature()) {
+        lightbulbService.getCharacteristic(hap.Characteristic.ColorTemperature)
           .on(CharacteristicEventTypes.GET, async (callback: CharacteristicGetCallback) => {
             callback(undefined, this.accStates.ColorTemperature);
-            this.lightbulbUpdateColorTemperature();
+            this.lightbulbAccessory.lightbulbUpdateColorTemperature();
           })
           .on(CharacteristicEventTypes.SET, async (value: CharacteristicValue, callback: CharacteristicSetCallback) => {
-            this.lightbulbSetColorTemperature(value);
+            this.lightbulbAccessory.lightbulbSetColorTemperature(value);
             callback();
           });
       }
 
-      if (this.config.updateIntervall) {
-        setInterval(() => {
-          if (this.lightbulbCheckOnOff()) {
-            this.lightbulbUpdateOn();
-          }
-          if (this.lightbulbCheckBrightness()) {
-            this.lightbulbUpdateBrightness();
-          }
-          if (this.lightbulbCheckHue()) {
-            this.lightbulbUpdateHue();
-          }
-          if (this.lightbulbCheckColorTemperature()) {
-            this.lightbulbUpdateColorTemperature();
-          }
-        }, this.config.updateIntervall);
-      }
+      this.serviceToExpose                     = lightbulbService;
+      this.lightbulbAccessory.lightbulbService = lightbulbService 
 
     }
 
     if (this.config.type == "switch") {
 
-      this.switchErrorCheck();
-      this.accessoryService = new hap.Service.Switch(this.name);
+      let switchService: Service = new hap.Service.Switch(this.name);
 
-      if (this.switchCheckOnOff()) {
-        this.accessoryService.getCharacteristic(hap.Characteristic.On)
+      this.switchAccessory.switchErrorCheck();
+
+      if (this.switchAccessory.switchCheckOnOff()) {
+        switchService.getCharacteristic(hap.Characteristic.On)
           .on(CharacteristicEventTypes.GET, async (callback: CharacteristicGetCallback) => {
             callback(undefined, this.accStates.On);
-            this.switchUpdateOn();
+            this.switchAccessory.switchUpdateOn();
           })
           .on(CharacteristicEventTypes.SET, async (value: CharacteristicValue, callback: CharacteristicSetCallback) => {
-            this.switchSetOn(value);
+            this.switchAccessory.switchSetOn(value);
             callback();
           });
       }
 
-      if (this.config.updateIntervall) {
-        setInterval(() => {
-          if (this.switchCheckOnOff()) {
-            this.switchUpdateOn();
-          }
-        }, this.config.updateIntervall);
-      }
+      this.serviceToExpose               = switchService;
+      this.switchAccessory.switchService = switchService 
 
     }
 
     if (this.config.type == "outlet") {
 
-      this.outletErrorCheck();
-      this.accessoryService = new hap.Service.Outlet(this.name);
+      let outletService: Service = new hap.Service.Outlet(this.name);
 
-      if (this.outletCheckOnOff()) {
-        this.accessoryService.getCharacteristic(hap.Characteristic.On)
+      this.outletAccessory.outletErrorCheck();
+
+      if (this.outletAccessory.outletCheckOnOff()) {
+        outletService.getCharacteristic(hap.Characteristic.On)
           .on(CharacteristicEventTypes.GET, async (callback: CharacteristicGetCallback) => {
             callback(undefined, this.accStates.On);
-            this.outletUpdateOn();
+            this.outletAccessory.outletUpdateOn();
           })
           .on(CharacteristicEventTypes.SET, async (value: CharacteristicValue, callback: CharacteristicSetCallback) => {
-            this.outletSetOn(value);
+            this.outletAccessory.outletSetOn(value);
             callback();
           });
       }
 
-      if (this.config.updateIntervall) {
-        setInterval(() => {
-          if (this.outletCheckOnOff()) {
-            this.outletUpdateOn();
-          }
-        }, this.config.updateIntervall);
-      }
+      this.serviceToExpose               = outletService;
+      this.outletAccessory.outletService = outletService 
 
     }
 
@@ -180,440 +171,14 @@ class HttpIotAccessory implements AccessoryPlugin {
         .setCharacteristic(hap.Characteristic.FirmwareRevision, pjson.version)
         .setCharacteristic(hap.Characteristic.SerialNumber, md5(this.name + pjson.model));
 
-    // log.info("Lightbulb finished initializing!");
-  }
-
-  lightbulbErrorCheck() {
-    if (this.config.lightbulbGetOn && this.config.lightbulbSetOn && this.config.lightbulbSetOff) {
-      if (!this.config.lightbulbGetOn.url || !this.config.lightbulbGetOn.method || !this.config.lightbulbGetOn.pattern || 
-        !this.config.lightbulbSetOn.url || !this.config.lightbulbSetOn.method || !this.config.lightbulbSetOff.url || !this.config.lightbulbSetOff.method) {
-        this.log.error('Config is not correct!');
-      }
-    } else {
-      this.log.error('Config is not correct!');
-    }
-  }
-  lightbulbCheckOnOff(): boolean {
-    let check = true;
-    if (!this.config.lightbulbGetOn || !this.config.lightbulbSetOn || !this.config.lightbulbSetOff ) {
-      check = false;
-    }
-    return check;
-  }
-  lightbulbCheckBrightness(): boolean {
-    let check = true;
-    if (!this.config.lightbulbGetBrightness || !this.config.lightbulbSetBrightness ) {
-      check = false;
-    }
-    return check;
-  }
-  lightbulbCheckHue(): boolean {
-    let check = true;
-    if (!this.config.lightbulbGetHue || !this.config.lightbulbSetHue ) {
-      check = false;
-    }
-    return check;
-  }
-  lightbulbCheckColorTemperature(): boolean {
-    let check = true;
-    if (!this.config.lightbulbGetColorTemperature || !this.config.lightbulbSetColorTemperature ) {
-      check = false;
-    }
-    return check;
-  }
-
-  switchErrorCheck() {
-    if (this.config.switchGetOn && this.config.switchSetOn && this.config.switchSetOff) {
-      if (!this.config.switchGetOn.url || !this.config.switchGetOn.method || !this.config.switchGetOn.pattern || 
-        !this.config.switchSetOn.url || !this.config.switchSetOn.method || !this.config.switchSetOff.url || !this.config.switchSetOff.method) {
-        this.log.error('Config is not correct!');
-      }
-    } else {
-      this.log.error('Config is not correct!');
-    }
-  }
-  switchCheckOnOff(): boolean {
-    let check = true;
-    if (!this.config.switchGetOn || !this.config.switchSetOn || !this.config.switchSetOff ) {
-      check = false;
-    }
-    return check;
-  }
-
-  outletErrorCheck() {
-    if (this.config.outletGetOn && this.config.outletSetOn && this.config.outletSetOff) {
-      if (!this.config.outletGetOn.url || !this.config.outletGetOn.method || !this.config.outletGetOn.pattern || 
-        !this.config.outletSetOn.url || !this.config.outletSetOn.method || !this.config.outletSetOff.url || !this.config.outletSetOff.method) {
-        this.log.error('Config is not correct!');
-      }
-    } else {
-      this.log.error('Config is not correct!');
-    }
-  }
-  outletCheckOnOff(): boolean {
-    let check = true;
-    if (!this.config.outletGetOn || !this.config.outletSetOn || !this.config.outletSetOff ) {
-      check = false;
-    }
-    return check;
   }
 
   identify(): void {
-    this.log("Lightbulb Identify!");
+    this.log("HTTP IoT Identify!");
   }
 
   getServices(): Service[] {
-    return [
-      this.informationService,
-      this.accessoryService,
-    ];
-  }
-
-  async lightbulbUpdateOn() {
-    axios({
-      method: this.config.lightbulbGetOn.method.toLowerCase(),
-      url: this.config.lightbulbGetOn.url
-    })
-      .then( (response: any) => {
-        // this.log(response.data.trim());
-        let on = false;
-        if (response.data.trim() == this.config.lightbulbGetOn.pattern) {
-          on = true;
-        }
-        this.accStates.On = on;
-        if (this.config.debugMsgLog) {
-          this.log.info("Current On is: " + (this.accStates.On ? "ON": "OFF"));
-        }
-        this.accessoryService.updateCharacteristic(this.api.hap.Characteristic.On, on);
-      })
-      .catch( (error: any) => {
-        if (this.config.debugMsgLog) {
-          this.log.info("Error in Get On: %s", error.message);
-        }
-      });
-  }
-
-  async lightbulbSetOn(value: CharacteristicValue) {
-    this.accStates.On = value as boolean;
-    let url: String;
-    let method: String;
-    if (value) {
-      url    = this.config.lightbulbSetOn.url;
-      method = this.config.lightbulbSetOn.method.toLowerCase(); 
-    } else {
-      url    = this.config.lightbulbSetOff.url;
-      method = this.config.lightbulbSetOff.method.toLowerCase(); 
-    }
-    axios({
-      method: method,
-      url: url
-    })
-      .then( (response: any) => {
-        // this.log(response.data.trim());
-        if (this.config.debugMsgLog) {
-          this.log.info("On was set to: " + (this.accStates.On ? "ON": "OFF"));
-        }
-      })
-      .catch( (error: any) => {
-        if (this.config.debugMsgLog) {
-          this.log.info("Error in Set On: %s", error.message);
-        }
-      });
-  }
-
-  async lightbulbUpdateBrightness() {
-    axios({
-      method: this.config.lightbulbGetBrightness.method.toLowerCase(),
-      url: this.config.lightbulbGetBrightness.url
-    })
-      .then( (response: any) => {
-        // this.log(response.data.trim());
-        let brightnessString: string = response.data.trim();
-        if (this.config.lightbulbGetBrightness.removeBefore) {
-          brightnessString = brightnessString.replace(this.config.lightbulbGetBrightness.removeBefore, "");
-        }
-        if (this.config.lightbulbGetBrightness.removeAfter) {
-          brightnessString = brightnessString.replace(this.config.lightbulbGetBrightness.removeAfter, "");
-        }
-        const newBrightness: number = parseInt(brightnessString.trim(), 10);
-        if (newBrightness >= 0 && newBrightness <= 100) {
-          this.accStates.Brightness = newBrightness;
-          if (this.config.debugMsgLog) {
-            this.log.info("Brightness is: %i", this.accStates.Brightness);
-          }
-          this.accessoryService.updateCharacteristic(this.api.hap.Characteristic.Brightness, this.accStates.Brightness);
-        } else {
-          if (this.config.debugMsgLog) {
-            this.log.info("Current Brightness is not correct: '%s'! [0 - 100]", brightnessString.trim());
-          }
-        }
-      })
-      .catch( (error: any) => {
-        if (this.config.debugMsgLog) {
-          this.log.info("Error in Get Brightness: %s", error.message);
-        }
-      });
-  }
-
-  async lightbulbSetBrightness(value: CharacteristicValue) {
-    this.accStates.Brightness = value as number;
-    let url: string    = this.config.lightbulbSetBrightness.url;
-    if (this.config.lightbulbSetBrightness.replaceNumber) {
-      url = url.replace(this.config.lightbulbSetBrightness.replaceNumber, this.accStates.Brightness.toString());
-    } else {
-      url = url + this.accStates.Brightness.toString();
-    }
-    // this.log.info(url);
-    axios({
-      method: this.config.lightbulbSetBrightness.method.toLowerCase(),
-      url: url
-    })
-      .then( (response: any) => {
-        // this.log(response.data.trim());
-        if (this.config.debugMsgLog) {
-          this.log.info("Brightness was set to: %i", this.accStates.Brightness);
-        }
-      })
-      .catch( (error: any) => {
-        if (this.config.debugMsgLog) {
-          this.log.info("Error in Set Brightness: %s", error.message);
-        }
-      });
-  }
-
-  async lightbulbUpdateHue() {
-    axios({
-      method: this.config.lightbulbGetHue.method.toLowerCase(),
-      url: this.config.lightbulbGetHue.url
-    })
-      .then( (response: any) => {
-        // this.log(response.data.trim());
-        let hueString: string = response.data.trim();
-        if (this.config.lightbulbGetHue.removeBefore) {
-          hueString = hueString.replace(this.config.lightbulbGetHue.removeBefore, "");
-        }
-        if (this.config.lightbulbGetHue.removeAfter) {
-          hueString = hueString.replace(this.config.lightbulbGetHue.removeAfter, "");
-        }
-        const newHue: number = parseInt(hueString.trim(), 10);
-        if (newHue >= 0 && newHue <= 360) {
-          this.accStates.Hue = newHue;
-          if (this.config.debugMsgLog) {
-            this.log.info("Current Hue is: %i", this.accStates.Hue);
-          }
-          this.accessoryService.updateCharacteristic(this.api.hap.Characteristic.Hue, this.accStates.Hue);
-        } else {
-          if (this.config.debugMsgLog) {
-            this.log.info("Current Hue is not correct: '%s' (>> %i)! [0 - 360]", hueString.trim(), newHue);
-          }
-        }
-      })
-      .catch( (error: any) => {
-        if (this.config.debugMsgLog) {
-          this.log.info("Error in Get Hue: %s", error.message);
-        }
-      });
-  }
-
-  async lightbulbSetHue(value: CharacteristicValue) {
-    this.accStates.Hue = value as number;
-    let url: string    = this.config.lightbulbSetHue.url;
-    if (this.config.lightbulbSetHue.replaceNumber) {
-      url = url.replace(this.config.lightbulbSetHue.replaceNumber, this.accStates.Hue.toString());
-    } else {
-      url = url + this.accStates.Hue.toString();
-    }
-    // this.log.info(url);
-    axios({
-      method: this.config.lightbulbSetHue.method.toLowerCase(),
-      url: url
-    })
-      .then( (response: any) => {
-        // this.log(response.data.trim());
-        if (this.config.debugMsgLog) {
-          this.log.info("Lightbulb state hue was set to: %i", this.accStates.Hue);
-        }
-      })
-      .catch( (error: any) => {
-        if (this.config.debugMsgLog) {
-          this.log.info("Error in Set Hue: %s", error.message);
-        }
-      });
-  }
-
-  async lightbulbUpdateColorTemperature() {
-    axios({
-      method: this.config.lightbulbGetColorTemperature.method.toLowerCase(),
-      url: this.config.lightbulbGetColorTemperature.url
-    })
-      .then( (response: any) => {
-        // this.log(response.data.trim());
-        let colorTemperatureString: string = response.data.trim();
-        if (this.config.lightbulbGetColorTemperature.removeBefore) {
-          colorTemperatureString = colorTemperatureString.replace(this.config.lightbulbGetColorTemperature.removeBefore, "");
-        }
-        if (this.config.lightbulbGetColorTemperature.removeAfter) {
-          colorTemperatureString = colorTemperatureString.replace(this.config.lightbulbGetColorTemperature.removeAfter, "");
-        }
-        let newColorTemperature: number = parseInt(colorTemperatureString.trim(), 10);
-        if (this.config.lightbulbGetColorTemperature.unit == "kelvin") {
-          if (newColorTemperature > 0) {
-            newColorTemperature = (1000000 / newColorTemperature); // mired = 1000000 / kelvin
-          }
-        }
-        if (newColorTemperature >= 140 && newColorTemperature <= 500) {
-          this.accStates.ColorTemperature = newColorTemperature;
-          if (this.config.debugMsgLog) {
-            this.log.info("Current Color Temperature is: %i", this.accStates.ColorTemperature);
-          }
-          this.accessoryService.updateCharacteristic(this.api.hap.Characteristic.ColorTemperature, this.accStates.ColorTemperature);
-        } else {
-          if (this.config.debugMsgLog) {
-            this.log.info("Current Color Temperature is not correct: '%s' (>> %i mired)! [140 - 500 mired] [mired = 1000000 / kelvin]", colorTemperatureString.trim(), newColorTemperature);
-          }
-        }
-      })
-      .catch( (error: any) => {
-        if (this.config.debugMsgLog) {
-          this.log.info("Error in Get Color Temperature: %s", error.message);
-        }
-      });
-  }
-
-  async lightbulbSetColorTemperature(value: CharacteristicValue) {
-    this.accStates.ColorTemperature = value as number;
-    let newColorTemperature: number = this.accStates.ColorTemperature;
-    if (this.config.lightbulbSetColorTemperature.unit == "kelvin") {
-      if (newColorTemperature > 0) {
-        newColorTemperature = (1000000 / newColorTemperature); // kelvin = 1000000 / mired
-      }
-    }
-    let url: string    = this.config.lightbulbSetColorTemperature.url;
-    if (this.config.lightbulbSetColorTemperature.replaceNumber) {
-      url = url.replace(this.config.lightbulbSetColorTemperature.replaceNumber, newColorTemperature.toString());
-    } else {
-      url = url + newColorTemperature.toString();
-    }
-    // this.log.info(url);
-    axios({
-      method: this.config.lightbulbSetColorTemperature.method.toLowerCase(),
-      url: url
-    })
-      .then( (response: any) => {
-        // this.log(response.data.trim());
-        if (this.config.debugMsgLog) {
-          this.log.info("Color Temperature was set to: %i", this.accStates.ColorTemperature);
-        }
-      })
-      .catch( (error: any) => {
-        if (this.config.debugMsgLog) {
-          this.log.info("Error in Set Color Temperature: %s", error.message);
-        }
-      });
-  }
-
-  async lightbulbUpdateOn() {
-    axios({
-      method: this.config.lightbulbGetOn.method.toLowerCase(),
-      url: this.config.lightbulbGetOn.url
-    })
-      .then( (response: any) => {
-        // this.log(response.data.trim());
-        let on = false;
-        if (response.data.trim() == this.config.lightbulbGetOn.pattern) {
-          on = true;
-        }
-        this.accStates.On = on;
-        if (this.config.debugMsgLog) {
-          this.log.info("Current On is: " + (this.accStates.On ? "ON": "OFF"));
-        }
-        this.accessoryService.updateCharacteristic(this.api.hap.Characteristic.On, on);
-      })
-      .catch( (error: any) => {
-        if (this.config.debugMsgLog) {
-          this.log.info("Error in Get On: %s", error.message);
-        }
-      });
-  }
-
-  async switchSetOn(value: CharacteristicValue) {
-    this.accStates.On = value as boolean;
-    let url: String;
-    let method: String;
-    if (value) {
-      url    = this.config.switchSetOn.url;
-      method = this.config.switchSetOn.method.toLowerCase(); 
-    } else {
-      url    = this.config.switchSetOff.url;
-      method = this.config.switchSetOff.method.toLowerCase(); 
-    }
-    axios({
-      method: method,
-      url: url
-    })
-      .then( (response: any) => {
-        // this.log(response.data.trim());
-        if (this.config.debugMsgLog) {
-          this.log.info("On was set to: " + (this.accStates.On ? "ON": "OFF"));
-        }
-      })
-      .catch( (error: any) => {
-        if (this.config.debugMsgLog) {
-          this.log.info("Error in Set On: %s", error.message);
-        }
-      });
-  }
-
-  async switchUpdateOn() {
-    axios({
-      method: this.config.switchGetOn.method.toLowerCase(),
-      url: this.config.switchGetOn.url
-    })
-      .then( (response: any) => {
-        // this.log(response.data.trim());
-        let on = false;
-        if (response.data.trim() == this.config.switchGetOn.pattern) {
-          on = true;
-        }
-        this.accStates.On = on;
-        if (this.config.debugMsgLog) {
-          this.log.info("Current On is: " + (this.accStates.On ? "ON": "OFF"));
-        }
-        this.accessoryService.updateCharacteristic(this.api.hap.Characteristic.On, on);
-      })
-      .catch( (error: any) => {
-        if (this.config.debugMsgLog) {
-          this.log.info("Error in Get On: %s", error.message);
-        }
-      });
-  }
-
-  async outletSetOn(value: CharacteristicValue) {
-    this.accStates.On = value as boolean;
-    let url: String;
-    let method: String;
-    if (value) {
-      url    = this.config.outletSetOn.url;
-      method = this.config.outletSetOn.method.toLowerCase(); 
-    } else {
-      url    = this.config.outletSetOff.url;
-      method = this.config.outletSetOff.method.toLowerCase(); 
-    }
-    axios({
-      method: method,
-      url: url
-    })
-      .then( (response: any) => {
-        // this.log(response.data.trim());
-        if (this.config.debugMsgLog) {
-          this.log.info("On was set to: " + (this.accStates.On ? "ON": "OFF"));
-        }
-      })
-      .catch( (error: any) => {
-        if (this.config.debugMsgLog) {
-          this.log.info("Error in Set On: %s", error.message);
-        }
-      });
+    return [ this.informationService, this.serviceToExpose ];
   }
 
 }
